@arsenm 2025-09-06
> After this patch: 0.5574s

Maybe this should be using small sets? I wouldn't expect there to be many duplicates to resolve